all form templates are created and used
