+++ package.json
@@ -10,6 +10,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.3.2",
     "@testing-library/user-event": "^7.1.2",
+    "add": "^2.0.6",
     "apollo-boost": "^0.4.7",
     "apollo-cache-inmemory": "^1.6.5",
     "apollo-link-http": "^1.5.16",
@@ -18,6 +19,7 @@
     "axios": "^0.19.2",
     "connected-react-router": "^6.8.0",
     "dotenv": "^8.2.0",
+    "git": "^0.1.5",
     "graphql": "^14.6.0",
     "history": "^4.10.1",
     "lodash.flowright": "^3.5.0",
