+++ src/components/formFlow/TechSkills.js
@@ -7,6 +7,8 @@ import {
   removeTechSkill,
 } from "../../actions/resumeFormActions.js";
 
+import SingleFieldFormTemplate from "./formsTemplate/singleFieldFormTemplate"
+
 import {
   Avatar,
   Button,
@@ -194,26 +196,7 @@ function TechSkills(props) {
               Postgres, SQLite, Knex,etc)
             </Typography>
             <form className={classes.form} onSubmit={anotherSkill}>
-              <Grid className={classes.formContainer} fullWidth>
-                <TextField
-                  variant="outlined"
-                  margin="normal"
-                  className={classes.skillTextField}
-                  fullWidth
-                  name="skill"
-                  label="Tech Skill"
-                  id="skill"
-                  onChange={onChange}
-                  value={info.skill}
-                />
-                <AddBoxIcon
-                  fullWidth
-                  className={classes.addIcon}
-                  color="primary"
-                  onClick={anotherSkill}
-                />
-              </Grid>
-
+              <SingleFieldFormTemplate onChange={onChange} info={info.skill} anotherOne={anotherSkill} name="skill" label="Tech Skill" />
               <Grid className={classes.skillContainer}>
                 <Paper
                   component="ul"
