+++ src/components/formFlow/reviewForm/educationCard.js
@@ -5,6 +5,8 @@ import "../formStyles/reviewForm.css";
 //Icon import
 import EditIcon from "@material-ui/icons/Edit";
 
+import EducationFormTemplate from "../formsTemplate/educationFormTemplate"
+
 import {
   CardContent,
   makeStyles,
@@ -74,79 +76,7 @@ function EducationCard(props) {
     return (
       <CardContent>
         <form onSubmit={saveInfo}>
-          <FormControl className={classes.selectorForm}>
-            <InputLabel className={classes.selectorText} id="type">
-              Education
-            </InputLabel>
-            <Select
-              className={classes.textField}
-              variant="outlined"
-              fullWidth
-              id="type"
-              label="Education"
-              name="type"
-              autoFocus
-              onChange={onChange}
-              value={info.type}
-            >
-              <MenuItem value={"College"}>College</MenuItem>
-              <MenuItem value={"University"}>University</MenuItem>
-              <MenuItem value={"Certification"}>Certification</MenuItem>
-              <MenuItem value={"Course"}>Course</MenuItem>
-            </Select>
-          </FormControl>
-          <TextField
-            variant="outlined"
-            margin="normal"
-            required
-            fullWidth
-            name="schoolName"
-            label="Name of the school"
-            id="schoolName"
-            onChange={onChange}
-            value={info.schoolName}
-          />
-          <TextField
-            variant="outlined"
-            margin="normal"
-            required
-            fullWidth
-            name="yearIn"
-            type="date"
-            label="School Starting Date"
-            id="yearIn"
-            InputLabelProps={{
-              shrink: true,
-            }}
-            onChange={onChange}
-            value={info.yearIn}
-          />
-          <TextField
-            variant="outlined"
-            margin="normal"
-            required
-            fullWidth
-            type="date"
-            name="yearOut"
-            label="Finishing School"
-            id="yearOut"
-            InputLabelProps={{
-              shrink: true,
-            }}
-            onChange={onChange}
-            value={info.yearOut}
-          />
-          <TextField
-            variant="outlined"
-            margin="normal"
-            required
-            fullWidth
-            name="certificateName"
-            label="Name of the certificate"
-            id="certificateName"
-            onChange={onChange}
-            value={info.certificateName}
-          />
+          <EducationFormTemplate info={info} onChange={onChange} />
           <Button
             type="submit"
             fullWidth
