+++ src/components/formFlow/reviewForm/projectCard.js
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 
 import "../formStyles/reviewForm.css";
+import ProjectsFormTemplate from "../formsTemplate/projectsFormTemplate"
 
 //Icon import
 import EditIcon from "@material-ui/icons/Edit";
@@ -66,76 +67,7 @@ function ProjectCard(props) {
     return (
       <CardContent>
         <form onSubmit={saveInfo}>
-          <TextField
-            variant="outlined"
-            margin="normal"
-            required
-            fullWidth
-            name="projectName"
-            label="Name of the Project"
-            id="projectName"
-            onChange={onChange}
-            value={info.projectName}
-          />
-          <TextField
-            variant="outlined"
-            margin="normal"
-            fullWidth
-            name="projectStartDate"
-            type="date"
-            label="Starting Date (Optional)"
-            id="projectStartDate"
-            InputLabelProps={{
-              shrink: true,
-            }}
-            onChange={onChange}
-            value={info.projectStartDate}
-          />
-          <TextField
-            variant="outlined"
-            margin="normal"
-            fullWidth
-            type="date"
-            name="projectEndDate"
-            label="End Date (Optional)"
-            id="projectEndDate"
-            InputLabelProps={{
-              shrink: true,
-            }}
-            onChange={onChange}
-            value={info.projectEndDate}
-          />
-          <TextField
-            variant="outlined"
-            margin="normal"
-            fullWidth
-            name="role"
-            label="Role"
-            id="role"
-            onChange={onChange}
-            value={info.role}
-          />
-          <TextField
-            variant="outlined"
-            margin="normal"
-            required
-            fullWidth
-            name="roleDescription"
-            label="Role Description"
-            id="roleDescription"
-            onChange={onChange}
-            value={info.roleDescription}
-          />
-          <TextField
-            variant="outlined"
-            margin="normal"
-            fullWidth
-            name="link"
-            label="Link to Project"
-            id="link"
-            onChange={onChange}
-            value={info.link}
-          />
+        <ProjectsFormTemplate info={info} onChange={onChange} />
           <Button
             type="submit"
             fullWidth
