+++ src/components/formFlow/workHistory.js
@@ -6,6 +6,8 @@ import { addWorkData, updateWorkData } from "../../actions/resumeFormActions.js"
 
 import JobHistoryCard from "./reviewForm/jobHistoryCard"
 
+import WorkHistoryFormTemplate from "./formsTemplate/workHistoryFormTemplate"
+
 import {
   Avatar,
   Button,
@@ -149,64 +151,7 @@ function WorkHistory(props) {
         <Grid item xs={12} sm={8} md={9} component={Paper} elevation={6} square>
           <div className={classes.paper}>
             <form id="workForm" className={classes.form} onSubmit={nextPage}>
-              <TextField
-                variant="outlined"
-                margin="normal"
-                fullWidth
-                name="jobTitle"
-                label="Job Title"
-                id="jobTitle"
-                onChange={onChange}
-                value={info.jobTitle}
-              />
-              <TextField
-                variant="outlined"
-                margin="normal"
-                fullWidth
-                name="companyName"
-                label="Name of the company"
-                id="companyName"
-                onChange={onChange}
-                value={info.companyName}
-              />
-              <TextField
-                variant="outlined"
-                margin="normal"
-                fullWidth
-                type="date"
-                name="startYear"
-                label="Starting Date"
-                id="startYear"
-                InputLabelProps={{
-                  shrink: true,
-                }}
-                onChange={onChange}
-                value={info.startYear}
-              />
-              <TextField
-                variant="outlined"
-                margin="normal"
-                fullWidth
-                name="endYear"
-                label="End Date"
-                type="date"
-                InputLabelProps={{
-                  shrink: true,
-                }}
-                id="endYear"
-                onChange={onChange}
-                value={info.endYear}
-              />
-              <TextField
-                variant="outlined"
-                margin="normal"
-                fullWidth
-                name="jobDescription"
-                label="Job Description"
-                id="jobDescription"
-                onChange={onChange}
-                value={info.jobDescription}
-              />
+              <WorkHistoryFormTemplate info={info} onChange={onChange} />
               <Button
                 type="reset"
                 fullWidth
